feat: get offers from all pages
Uses a timeout instead of a more reliable assertion when it comes to
moving to next pages. This is unreliable, but works for the time being.

It would be nice to use a Cypress-like approach in the near future, and
run assertions on a given interval, with some timeout, to allow some
asynchronous actions to finish.