fix: add port disconnection handling
Ports that cannot connect (perhaps because the receiver is not set up
yet after a page reload) will emit an `onDisconnect` event that needs to
be handled by the extension. Handling in this case means inspecting the
`chrome.runtime.lastError` property in the `onDisconnect` handler.
Otherwise, some next operation on the port will throw an error.

The error handling needs to be made more robust and applied to every
port. This is just a basic implementation for the most important case.