fix: catch all ready events from tabs
A "ready" event could have been missed if it fired between making an
update to the tab and setting up an event listener for that event.

Setting up the event listener first and then only waiting for the result
solves the problem.

It leverages the native Promise functionality of resolving with the value
immediately if the promise is already in the "resolved" state.