fix: bundle all code for content scripts and background page
Creating common chunks for code shared between multiple content scripts
and the background page is not possible, becuase importing code in those
files is not allowed. Each chunk needs to have all the code it requires.

Also, the service-worker.js script is now at the root of the build
directory.