feature: display and cache the offers list
The offers list is cached in the service worker. It is sent to the popup
immediately when it is rendered, meaning that it can reuse a previous
list of offers without needing to do a scan.

The popup also displays the offers as a list, instead of just
stringifying JSON.