chore: upgrade snowpack to 3.8.8
The upgrade introduced a regression related to type-only files.
They were not included in the final build, but they were still imported
with `export * from ...` exports.

It is discussed in
https://github.com/withastro/snowpack/discussions/1589#discussioncomment-130176

The fix is to add some JS to that file (e.g. `export {}`). This way the
file stays in the build.